requirements.txt
anthropic>=0.25.0
google-generativeai>=0.5.0
openai>=1.30.0
requests>=2.31.0
Pillow>=10.0.0
jinja2>=3.1.0
python-dotenv>=1.0.0
supabase>=2.0.0
paramiko>=3.0.0
